removed unneeded tiledlayers list, form, controller
