+++ publisher-web/app/controllers/GroupsLayersCommon.java
@@ -1,5 +1,10 @@
 package controllers;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+import nl.idgis.publisher.domain.web.TiledLayer;
 import play.Play;
 import play.mvc.Controller;
 import play.mvc.Security;
