--- publisher-web/app/views/tiledlayers/form.scala.html
@@ -1,34 +1,0 @@
-@(tiledlayerForm: Form[controllers.Tiledlayers.TiledLayerForm], create: Boolean)
-
-@import nl.idgis.publisher.domain.web._ 
-@layout.application( title=Domain.message("web.application.page.tiledlayers.form.title")){
-
-<div class="page-content">
-	@if(create){
-		<h1 class="page-header">@Domain.message("web.application.page.tiledlayers.form.header.new")</h1>
-	}else{
-		<h1 class="page-header">@Domain.message("web.application.page.tiledlayers.form.header.edit", tiledlayerForm("name").value) </h1>
-	}
-
-	@helper.formerrors(tiledlayerForm, Domain.message("web.application.page.tiledlayers.name").toLowerCase())	
-
-	
-	<div class="row">
-		<div class="col-lg-12">
-			<form class="form-horizontal" method="post" action="@routes.Tiledlayers.submitCreateUpdate()">
-				<input type="hidden" class="form-control" id="input-id" name="id" value="@tiledlayerForm("id").value" >
-				@tiledLayerBody(tiledlayerForm);
-				<div class="form-group">
-					<div class="col-sm-offset-2 col-sm-10">
-						<button class="btn btn-success" type="submit">@Domain.message("web.application.button.save")</button>
-						<a href="@routes.Tiledlayers.list()" 
-							class="btn btn-success"> @Domain.message("web.application.button.cancel")
-						</a>
-					</div>
-				</div>
-				
-			</form>
-		</div>
-	</div>
-</div>
-}
