--- publisher-web/app/views/tiledlayers/list.scala.html
@@ -1,51 +1,0 @@
-@(tiledlayers: nl.idgis.publisher.domain.response.Page[nl.idgis.publisher.domain.web.TiledLayer])
-
-@import nl.idgis.publisher.domain.web._ 
-@layout.application( title = Domain.message("web.application.page.tiledlayers.list.header") ) {
-
-<div class="page-content">
-	<h1 class="page-header">@Domain.message("web.application.page.tiledlayers.list.header")</h1>
-	
-	@* List of tiledlayers: *@
-	<div class="col-lg-9 ">
-		@if(tiledlayers.values.isEmpty) {
-		<div class="alert alert-info">@Domain.message("web.application.page.tiledlayers.list.empty")</div>
-		} else {
-		<div class="list-group" id="tiledlayer-list">
-
-			@for(tiledlayer <- tiledlayers.values) {
-			<div class="list-group-item js-tiledlayer-item" 
-				data-tiledlayer-id="@tiledlayer.id"
-				data-tiledlayer-name="@tiledlayer.name">
-				<div class="row">
-					<div class="col-md-10">
-						<h4 class="list-group-item-heading">
-							<a href="@routes.Tiledlayers.edit(tiledlayer.id)" class="text-warning">@tiledlayer.name</a>
-						</h4>
-					</div>
-					<div class="col-md-2">
-						<div class="btn-group pull-right">
-							<a href="@routes.Tiledlayers.edit(tiledlayer.id)"
-								class="btn btn-default" title=@Domain.message("web.application.button.edit")>
-								<span class="glyphicon glyphicon-edit"></span>
-							</a> 
-							<a href="@routes.Tiledlayers.delete(tiledlayer.id)" 
-								class="btn btn-warning" title=@Domain.message("web.application.button.remove")>
-								<span class="glyphicon glyphicon-remove"></span>
-							</a>
-						</div>
-					</div>
-				</div>
-			</div>
-			}
-		</div>
-		}
-		<div class="btn-group">
-			<a href="@routes.Tiledlayers.create()" 
-				class="btn btn-success" title=@Domain.message("web.application.button.add")>
-				<span class="glyphicon glyphicon-plus"></span> @Domain.message("web.application.button.add")
-			</a>
-		</div>
-	</div>
-</div>
-}
