+++ publisher-web/app/views/tiledlayers/tiledLayerBody.scala.html
@@ -1,4 +1,4 @@
-@(tiledlayerForm: Form[_ <: controllers.Tiledlayers.TiledLayerForm])
+@(tiledlayerForm: Form[_ <: controllers.TiledLayerForm])
 
 @import nl.idgis.publisher.domain.web._
