+++ publisher-web/conf/routes
@@ -73,13 +73,6 @@ POST	/styles/-/upload-file-raw						controllers.Styles.handleFileUploadRaw()
 GET		/styles/:styleId/edit							controllers.Styles.edit(styleId: String)
 POST 	/styles/:styleId/delete							controllers.Styles.delete(styleId : String)
 
-GET		/tiledlayers									controllers.Tiledlayers.list()
-GET 	/tiledlayers/-/create							controllers.Tiledlayers.create()
-POST 	/tiledlayers/-/createupdate						controllers.Tiledlayers.submitCreateUpdate()
-GET		/tiledlayers/:tiledlayerId/edit					controllers.Tiledlayers.edit(tiledlayerId: String)
-GET 	/tiledlayers/:tiledlayerId/delete				controllers.Tiledlayers.delete(tiledlayerId : String)
-
-
 # Logging
 GET		/logging/messages								controllers.Logging.messages()
 GET		/logging/messages/:logLevels/:page				controllers.Logging.messagesWithFilter(logLevels: String, page: Long)
